flask_mysql/belt_exam/car_dealership/requirements.txt
bcrypt==4.0.1
blinker==1.6.3
cffi==1.16.0
click==8.1.7
colorama==0.4.6
cryptography==41.0.4
Flask==3.0.0
Flask-Bcrypt==1.0.1
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
pycparser==2.21
PyMySQL==1.1.0
python-dateutil==2.8.2
six==1.16.0
Werkzeug==3.0.0
